+++ src/iota/addresses.h
@@ -5,7 +5,7 @@
 
 #define MAX_SECURITY 3
 
-void get_public_addr(const unsigned char *seed_bytes, uint32_t idx, uint8_t security,
-                   unsigned char *address_bytes);
+void get_public_addr(const unsigned char *seed_bytes, uint32_t idx,
+                     unsigned int security, unsigned char *address_bytes);
 
 #endif // ADDRESSES_H
